camera_based_model/medicine-verifier/requirements.txt
transformers>=4.40.0
torch>=2.2.0
Pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0
diskcache>=5.6.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-search-results>=2.4.0
